Update setup.cfg and tox.ini
Add Python 3.9 to the list of supported Pythons and replace
dashes (which are deprecated) with underscores.

Setuptools v54.1.0 introduces a warning that the use of dash-separated
options in 'setup.cfg' will not be supported in a future version [1].
Get ahead of the issue by replacing the dashes with underscores. Without
this, we see 'UserWarning' messages like the following on new enough
versions of setuptools:

  UserWarning: Usage of dash-separated 'description-file' will not be
  supported in future versions. Please use the underscore name
  'description_file' instead

[1] https://github.com/pypa/setuptools/commit/a2e9ae4cb

Change-Id: I12cc6206ebfe06b7cf3a50ea59bf1b55a3ff7303